Add project view, eventing and history

+++ app/Models/Project/Traits/Attribute/ProjectAttribute.php
@@ -70,7 +70,7 @@ trait ProjectAttribute
      */
     public function getDeleteButtonAttribute()
     {
-        if ($this->id != access()->id() && $this->id != 1) {
+        if (!$this->isApproved() && $this->id != access()->id() && $this->id != 1) {
             return '<a href="'.route('admin.projects.destroy', $this).'"
                  data-method="delete"
                  data-trans-button-cancel="'.trans('buttons.general.cancel').'"
